refactor: move create token to custom hook
