+++ components/Table/HeaderTable.tsx
@@ -1,110 +1,18 @@
 import { Header, Text, Button, Box, StyledOcticon } from "@primer/react";
 import { CheckIcon, SyncIcon } from "@primer/octicons-react";
-import {
-  createAssociatedTokenAccountInstruction,
-  createInitializeMintInstruction,
-  getMinimumBalanceForRentExemptMint,
-  MINT_SIZE,
-  TOKEN_PROGRAM_ID,
-} from "@solana/spl-token";
-import { useConnection, useWallet } from "@solana/wallet-adapter-react";
-import { Keypair, SystemProgram, Transaction } from "@solana/web3.js";
-import { RpcMethods } from "lib/spl";
-import { useRefresh, useSuccess, useHasMongoUri } from "contexts";
-import { useState } from "react";
+
+import { useWallet } from "@solana/wallet-adapter-react";
+import { useHasMongoUri } from "contexts";
 import { HourglassIcon } from "@primer/octicons-react";
-import { createMint } from "lib/create-token";
+import { useCreateToken } from "hooks";
 
 const HeaderTable: React.FC<{ tokensAmount: number }> = ({
   tokensAmount = 0,
 }) => {
-  const { r, refresh } = useRefresh();
-  const { connection } = useConnection();
-  const { publicKey, sendTransaction } = useWallet();
-  const { setMessage, setMint } = useSuccess();
-  const [minting, setMinting] = useState(false);
+  const { publicKey } = useWallet();
   const { hasMongoUri } = useHasMongoUri();
-  function triggerRefresh() {
-    refresh(!r);
-  }
-  async function createToken() {
-    if (!publicKey) {
-      try {
-        setMinting(true);
-        const tokenData = await createMint();
-        if (!tokenData) return;
-        await fetch("api/mongo-new-token", {
-          method: "POST",
-          body: JSON.stringify(tokenData),
-        });
-        setMinting(false);
-        setMessage(
-          `Success! New mint ${tokenData.token.slice(0, 12)}... created.`
-        );
-        setMint(tokenData.token);
+  const { createToken, minting, triggerRefresh } = useCreateToken();
 
-        triggerRefresh();
-        return;
-      } catch (e) {
-        setMinting(false);
-        console.error(e);
-        return;
-      }
-    }
-    try {
-      setMinting(true);
-      const mint = Keypair.generate();
-
-      const rpc = new RpcMethods(connection);
-
-      const ata = await rpc.getAssociatedTokenAccount(
-        mint.publicKey.toBase58(),
-        publicKey.toBase58()
-      );
-
-      let tx = new Transaction()
-        .add(
-          SystemProgram.createAccount({
-            fromPubkey: publicKey,
-            newAccountPubkey: mint.publicKey,
-            space: MINT_SIZE,
-            lamports: await getMinimumBalanceForRentExemptMint(connection),
-            programId: TOKEN_PROGRAM_ID,
-          })
-        )
-        .add(
-          createInitializeMintInstruction(
-            mint.publicKey,
-            0,
-            publicKey,
-            publicKey
-          )
-        )
-        .add(
-          createAssociatedTokenAccountInstruction(
-            publicKey,
-            ata,
-            publicKey,
-            mint.publicKey
-          )
-        );
-      tx.recentBlockhash = (await connection.getLatestBlockhash()).blockhash;
-      tx.feePayer = publicKey;
-      tx.sign(mint);
-      const signature = await sendTransaction(tx, connection);
-      await rpc.confirmTransaction(signature);
-      triggerRefresh();
-      setMessage(
-        `Success! New mint ${mint.publicKey
-          .toBase58()
-          .slice(0, 12)}... created.`
-      );
-      setMint(mint.publicKey.toBase58());
-    } catch (e) {
-      console.error(e);
-    }
-    setMinting(false);
-  }
   return (
     <Header
       style={{
