+++ components/Table/Row.tsx
@@ -10,7 +10,7 @@ import { IssueOpenedIcon, HourglassIcon } from "@primer/octicons-react";
 import { useWallet } from "@solana/wallet-adapter-react";
 import { useEffect } from "react";
 import { useMintAndTransfer } from "hooks";
-import { Actions, RowProps, Sources } from "types";
+import { Actions, RowProps } from "types";
 import { useRefresh } from "contexts";
 
 enum IssueColor {
