+++ hooks/index.ts
@@ -1,2 +1,3 @@
 export {default as useHandleDestroyAnimated} from './useHandleDestroyAnimated';
 export {default as useMintAndTransfer} from './useMintAndTransfer';
+export {default as useCreateToken} from './useCreateToken';
